Follow the RFC a bit more
Checking for at least HTTP 1.1 now. Next to that also checking if the
Upgrade header contains the "websocket" keyword and if the Connection
header is contains the "Upgrade" keyword. No longer saving the websocket
version either, as the RFC says I should just check if this is 13. Only
thing from the list in the RFC (RFC 6455 page 17 and 18) is the Host header.
But as we currently don't have a config option for this anyway I think
we can safely ignore this. And yes, I don't do anything special for point
10, 11 and 12 of that list. Simply because it isn't needed.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>