Implemented the length code 126 part of WebSocket
Didn't implement the 127 part yet and might not implement that at all,
this 126 part is slightly buggy anyway. As it'll just 'ignore' the entire
frame if it did read the length but can't read that amount of bytes yet.
Before I move on and even consider implementing the 127 code part I just
make it sort of persist in the websocket struct or something like that.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>