Implemented a FLOAT_RANGE type
FLOAT_RANGE type makes a bit more sense for things with for example
percentages. This is why the cpu module uses it from now on. Also
made a couple of changes to responses when loading, erroring etc.
The loaded message for example will now include the type of the module
and any extra data attached to it. Like for example with FLOAT_RANGE
the range (minimum/maximum value).

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>